Add pydantic models for Project and rename the Diagram model
