+++ besser/utilities/web_modeling_editor/backend/models/__init__.py
@@ -1,3 +1,4 @@
-from .class_diagram import ClassDiagramInput
+from .diagram import DiagramInput
+from .project import ProjectInput
 
-__all__ = ['ClassDiagramInput'] 
+__all__ = ['DiagramInput']
