+++ besser/utilities/web_modeling_editor/backend/services/__init__.py
@@ -1,5 +1,6 @@
 from .buml_to_json import parse_buml_content, domain_model_to_json
-from .json_to_buml import process_class_diagram, process_state_machine
+from .json_to_buml import *
 from .deploy_app import run_docker_compose
+from .json_to_buml_project import json_to_buml_project
 
 __all__ = ['parse_buml_content', 'domain_model_to_json', 'process_class_diagram', 'process_state_machine']
